Review processing reworked:  - AddReviewRequest class added, the class serves the purpose of receiving request for Review addition from client;  - DefaultReviewService.getReviewsByMovieId() eliminated, since it was not used;  - JdbcReviewDao.findAllByMovieId() eliminated since it was used only for testing purpose;  - uuid from the client passed as header, so the issue was fixed;  - rider-core dependency was removed since rider-spring dependency already contains it;   - application.properties has genre cache settings renamed in order to bring itto one naming convention;

+++ src/main/java/edu/ted/web/movieland/entity/Review.java
@@ -12,9 +12,9 @@ import java.time.LocalDate;
 public class Review {
     @Getter
     @Setter
-    private int reviewId;
+    private long reviewId;
     @Getter
-    private final int movieId;
+    private final long movieId;
     @Getter
     private final String text;
     @Getter
